will show meta content from the api response rather than that of the csv
in settings now parsing out meta content selector
in auction response also parsing out metaContentSpotsWithDetail

if NOT on mobile and api response gave us meta content, then it calls setupMetaContent to show tippies

updated prisma client type definition

setupMetaContent gets metaContentSelector and metaContentSpotsWithDetail
it uses the selector only to find prospective elements
it builds the tippy div using metaContent object.
it takes metaContent's heading to build first item with bold font weight
then splits metaContent's text on \n and builds the remaining paragraphs
it also adds a read more button if multiple items in the text

for each meta content spot returns it finds it in the dom.
dom is searched using the provided selector and then its text is used to look for match
the textContent of the dom element is used
textContent also gets all spaces and new line feeds first replaced with empty space as that is how we save them in the DB when the spot was built.