Fixing CLS for inline tooltip
inlineTooltip.css:
no longer has hidden class.
its visible class no longer ads height, instead just adds transition.
it builds tooltip with 0 opacity and visible class does transition on opacity to 1.
skip button has been replaced with close button

sendMessageToGa.ts has been re-written to not send page view events. instead the library is just wrapper for gtag will send whatever event and props that are given to it.

setupInlineTooltip.ts:
hidden class is not present when tooltip is built. this allows it to take the max-height its scrollable container requires.
skip has become close
when container is shown we just add visible class when not present. this does opacity transition.
handleClose event handler removes all our tooltips. it causes no CLS as after user interaction.
when page scroll observer detect item is visible is calls showAnswer, without checking if item is at the top of the page.
threshold to show answer is 0.5, no longer multiple thresholds as we no longer check location and bail.
build function addTooltip which does the insertion of the tooltip.
it first checks if the location of where it is inserting is eiher above or below the visible viewport. only then does it insert.
if item is in viewport, then it skips and adds event handler to call itself which will try inserting on scrollend event.
it inserts tooltip as append on the spot rather than after. this puts the tooltip inside the spot container.
it uses once on the scroll handler, so it is called only once to insert on scrollend event, where if it is unable to add it will just schedule another scrollend event listener.

setupMetaContent.ts:
calls ga library to save ga events with event name and props
on display it sets up bw_inline_tooltip_display prop with yes value
it no longer has any `await` on the calls to updateAuction, as this was making it wait and thus slowing down the time to which we call setupInlineTooltip and start the tooltip insertions
removed the initial default call to updateExtra with show_tippy
on AB test roll it sets bw_ab_test_result to yes and no.
added function loadTooltip which then calls setupInlineTooltip or setupHoverTooltip with correct params. then calling this function immediately or after DOMContentLoaded depending on the readyState of the page.